Move service now functions to shared
