--- packages/blocks/servicenow/src/lib/build-api-url.ts
@@ -1,9 +1,0 @@
-import { ServiceNowAuth } from './auth';
-
-export function buildServiceNowApiUrl(
-  auth: ServiceNowAuth,
-  path?: string,
-): string {
-  const baseUrl = `https://${auth.instanceName}.service-now.com/api/now/table`;
-  return path ? `${baseUrl}/${path}` : baseUrl;
-}
