--- packages/blocks/servicenow/src/lib/generate-auth-header.ts
@@ -1,13 +1,0 @@
-export const generateAuthHeader = ({
-  username,
-  password,
-}: {
-  username: string;
-  password: string;
-}) => {
-  return {
-    Authorization: `Basic ${Buffer.from(`${username}:${password}`).toString(
-      'base64',
-    )}`,
-  };
-};
